analyst_toolkit.ps1 initial upload.
This script will prompt for authentication for Microsoft / Azure AD. Once authenticated it will export logs (if there are entries found) for each user in the Microsoft tenant. Logs are drawn from Azure AD and Exchange Online so far. The first prompt for auth is from Azure AD and the second (after the azure logs are exported and azure ad session ended) is from Exchange Online. Much more to come on in terms of updates for more logs to pull and the second phase will be running a second script on the downloaded files found in the downloads folder and report relevant metrics per log.